LayoutTests/http/tests/security/contentSecurityPolicy/object-in-svg-foreignobject-expected.txt

CONSOLE MESSAGE: Refused to load plugin data from 'http://127.0.0.1:8000/plugins/resources/mock-plugin.pl' because it violates the following Content Security Policy directive: "object-src 'none'".

This test ensures that objects inside SVG foreignobject elements are beholden to the same policy as the rest of the document. This test passes if there is a console message saying the plugin was blocked.
